+++ app/utils/defaultSettings.js
@@ -68,6 +68,7 @@ export default {
   pauseBreaksFor5HoursShortcut: '',
   pauseBreaksUntilMorningShortcut: '',
   screen: 'primary',
+  breakContentScreen: 'all',
   trayIconStyle: 'default',
   currentTimeInBreaks: false,
   showTrayIcon: true,